# src/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIRIS EPM API
  description: |
    AIRIS Enterprise Performance Management (EPM) API provides comprehensive monitoring, 
    analytics, and performance optimization capabilities for enterprise systems.
    
    ## Features
    - Real-time performance monitoring
    - Data quality validation
    - Advanced analytics and reporting
    - Caching and optimization
    - Alert and notification system
    
    ## Authentication
    API uses Bearer token authentication. Include your API key in the Authorization header.
    
    ## Rate Limiting
    API requests are limited to 1000 requests per hour per API key.
    
  version: "1.0.0"
  contact:
    name: AIRIS EPM Support
    url: https://airis-epm.com/support
    email: support@airis-epm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.airis-epm.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /data-quality/anomalies:
    get:
      tags:
        - Data Quality
      summary: Get anomaly detection results
      description: Retrieve detected anomalies in the data
      parameters:
        - name: timeRange
          in: query
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
            default: 24h
          description: Time range for anomaly detection
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Minimum severity level
      responses:
        '200':
          description: Anomalies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomaliesResponse'

    post:
      tags:
        - Data Quality
      summary: Submit data for anomaly detection
      description: Submit data to be analyzed for anomalies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnomalyDetectionRequest'
      responses:
        '200':
          description: Anomaly detection completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyDetectionResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnomaliesResponse:
      type: object
      properties:
        success:
          type: boolean
        anomalies:
          type: array
          items:
            $ref: '#/components/schemas/Anomaly'
        total_count:
          type: integer
    Anomaly:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        type:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        description:
          type: string
        affected_metrics:
          type: array
          items:
            type: string
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
    AnomalyDetectionRequest:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            type: object
        algorithms:
          type: array
          items:
            type: string
            enum: [statistical, ml, hybrid]
        sensitivity:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
    AnomalyDetectionResponse:
      type: object
      properties:
        success:
          type: boolean
        anomalies:
          type: array
          items:
            $ref: '#/components/schemas/Anomaly'
        processing_time:
          type: number
...